lots of updates
made the edit csv data page (copied a lot of code cuz the dt objects
trunkate a lot of the data), backup creates a new file, made a lot of
the dt variables gobally accessible, you can view the contents of them
by just changing the comment on the editcsv page. also added styling to
datagrids.

i have a backup of rajaks commit if need be